+++ packages/import/package.js
@@ -34,7 +34,6 @@ Package.onUse(function(api) {
     'client/views/panels.html',
     'client/views/panels.js',
     'client/views/tabs.html',
-    'client/views/tabs.js',
     'client/router.js',   
   ],'client');