+++ packages/import/server/import.js
@@ -51,14 +51,24 @@ Meteor.methods({
 	importFile: function(fileObj, sets) {
 
 		if(!this.userId) return null;
+		
+		var mimes = [];
+		_.each(sets.mimeFileType, function(v,k) {
+			if(v===true)
+				mimes.push(k);
+		});
+		
+		if(!_.contains(mimes, fileObj.type)) {
+			console.log('Import: error ', _.omit(fileObj,'blob') );
+			throw new Meteor.Error(500, i18n('error_import_formatNotValid') );
+			return null;
+		}
 
 		var geo = JSON.parse(fileObj.blob),
 			importName = fileObj.name || K.Util.timeName(),
 			placeIds = [];
 
 		console.log('Import: file ', fileObj.name);
-
-	return ['ciao','bau','miao'];
 
 		if(geo && geo.features && geo.features.length>0) {
 
@@ -89,6 +99,6 @@ Meteor.methods({
 		else
 			console.log('Import: error json parse');
 
-		return placeIds;
+		return placeIds.length+' '+i18n('label_imported');
 	}
 });